chore: move Plant Growth hooks out of components dir [PT-185561938]
