+++ plant-growth/src/app/hooks/use-interactive-state.ts
@@ -1,6 +1,6 @@
 import { useCallback } from "react";
 import { useInteractiveState } from "@concord-consortium/lara-interactive-api";
-import { IInteractiveState } from "../../../types";
+import { IInteractiveState } from "../../types";
 
 export const useSaveInteractiveState = () => {
   const { setInteractiveState } = useInteractiveState<IInteractiveState>();
