+++ plant-growth/src/app/hooks/use-model-state.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from "@testing-library/react-hooks";
-import { defaultAuthoredState, InputAmount, OutputAmountValue } from "../../../types";
+import { defaultAuthoredState, InputAmount, OutputAmountValue } from "../../types";
 import { useModelState } from "./use-model-state";
 
 const testState = {
