+++ plant-growth/src/app/hooks/use-model-state.ts
@@ -1,5 +1,5 @@
 import { useCallback, useMemo, useState } from "react";
-import { defaultInitialState, IModelInputState, IModelOutputState } from "../../../types";
+import { defaultInitialState, IModelInputState, IModelOutputState } from "../../types";
 import { useSaveInteractiveState } from "./use-interactive-state";
 
 // initialOutputState can be either an object or a function of initialInputState which returns an object.
