+++ plant-growth/src/app/hooks/use-translation.scss
@@ -1,4 +1,4 @@
-@import "../vars.scss";
+@import "../components/vars.scss";
 
 .container {
   width: fit-content;
