+++ plant-growth/src/app/translations.tsx
@@ -1,46 +1,46 @@
-import NewTrial from "../assets/yugtun-audio/new-trial.mp3";
-import Controls from "../assets/yugtun-audio/controls.mp3";
-import Play from "../assets/yugtun-audio/play.mp3";
-import Day0 from "../assets/yugtun-audio/day-0.mp3";
-import Day4 from "../assets/yugtun-audio/day-4.mp3";
-import Day8 from "../assets/yugtun-audio/day-8.mp3";
-import Day12 from "../assets/yugtun-audio/day-12.mp3";
-import Day16 from "../assets/yugtun-audio/day-16.mp3";
-import Day20 from "../assets/yugtun-audio/day-20.mp3";
-import Day24 from "../assets/yugtun-audio/day-24.mp3";
-import Day28 from "../assets/yugtun-audio/day-28.mp3";
-import Co2 from "../assets/yugtun-audio/carbon-dioxide.mp3";
-import Light from "../assets/yugtun-audio/light.mp3";
-import Water from "../assets/yugtun-audio/water.mp3";
-import Full from "../assets/yugtun-audio/full.mp3";
-import Some from "../assets/yugtun-audio/some.mp3";
-import None from "../assets/yugtun-audio/none.mp3";
-import TrialsData from "../assets/yugtun-audio/trials-data.mp3";
-import Trial from "../assets/yugtun-audio/trial.mp3";
-import SugarCreated from "../assets/yugtun-audio/sugar-created.mp3";
-import SugarUsed from "../assets/yugtun-audio/sugar-used.mp3";
-import Low from "../assets/yugtun-audio/low.mp3";
-import Medium from "../assets/yugtun-audio/medium.mp3";
-import High from "../assets/yugtun-audio/high.mp3";
-import DeleteTrial from "../assets/yugtun-audio/delete-trial.mp3";
-import Graphs_XAxisLabel from "../assets/yugtun-audio/time-days.mp3";
-import Graphs_YAxisLabel from "../assets/yugtun-audio/amount.mp3";
-import Graphs_Trial1 from "../assets/yugtun-audio/trial-1-graphs.mp3";
-import Graphs_Trial2 from "../assets/yugtun-audio/trial-2-graphs.mp3";
-import Graphs_Trial3 from "../assets/yugtun-audio/trial-3-graphs.mp3";
-import Graphs_Trial4 from "../assets/yugtun-audio/trial-4-graphs.mp3";
-import Graphs_Trial5 from "../assets/yugtun-audio/trial-5-graphs.mp3";
-import Graphs_Trial6 from "../assets/yugtun-audio/trial-6-graphs.mp3";
-import Graphs_Trial7 from "../assets/yugtun-audio/trial-7-graphs.mp3";
-import Graphs_Trial8 from "../assets/yugtun-audio/trial-8-graphs.mp3";
-import Graphs_Trial9 from "../assets/yugtun-audio/trial-9-graphs.mp3";
+import NewTrial from "./assets/yugtun-audio/new-trial.mp3";
+import Controls from "./assets/yugtun-audio/controls.mp3";
+import Play from "./assets/yugtun-audio/play.mp3";
+import Day0 from "./assets/yugtun-audio/day-0.mp3";
+import Day4 from "./assets/yugtun-audio/day-4.mp3";
+import Day8 from "./assets/yugtun-audio/day-8.mp3";
+import Day12 from "./assets/yugtun-audio/day-12.mp3";
+import Day16 from "./assets/yugtun-audio/day-16.mp3";
+import Day20 from "./assets/yugtun-audio/day-20.mp3";
+import Day24 from "./assets/yugtun-audio/day-24.mp3";
+import Day28 from "./assets/yugtun-audio/day-28.mp3";
+import Co2 from "./assets/yugtun-audio/carbon-dioxide.mp3";
+import Light from "./assets/yugtun-audio/light.mp3";
+import Water from "./assets/yugtun-audio/water.mp3";
+import Full from "./assets/yugtun-audio/full.mp3";
+import Some from "./assets/yugtun-audio/some.mp3";
+import None from "./assets/yugtun-audio/none.mp3";
+import TrialsData from "./assets/yugtun-audio/trials-data.mp3";
+import Trial from "./assets/yugtun-audio/trial.mp3";
+import SugarCreated from "./assets/yugtun-audio/sugar-created.mp3";
+import SugarUsed from "./assets/yugtun-audio/sugar-used.mp3";
+import Low from "./assets/yugtun-audio/low.mp3";
+import Medium from "./assets/yugtun-audio/medium.mp3";
+import High from "./assets/yugtun-audio/high.mp3";
+import DeleteTrial from "./assets/yugtun-audio/delete-trial.mp3";
+import Graphs_XAxisLabel from "./assets/yugtun-audio/time-days.mp3";
+import Graphs_YAxisLabel from "./assets/yugtun-audio/amount.mp3";
+import Graphs_Trial1 from "./assets/yugtun-audio/trial-1-graphs.mp3";
+import Graphs_Trial2 from "./assets/yugtun-audio/trial-2-graphs.mp3";
+import Graphs_Trial3 from "./assets/yugtun-audio/trial-3-graphs.mp3";
+import Graphs_Trial4 from "./assets/yugtun-audio/trial-4-graphs.mp3";
+import Graphs_Trial5 from "./assets/yugtun-audio/trial-5-graphs.mp3";
+import Graphs_Trial6 from "./assets/yugtun-audio/trial-6-graphs.mp3";
+import Graphs_Trial7 from "./assets/yugtun-audio/trial-7-graphs.mp3";
+import Graphs_Trial8 from "./assets/yugtun-audio/trial-8-graphs.mp3";
+import Graphs_Trial9 from "./assets/yugtun-audio/trial-9-graphs.mp3";
 
-import InstructionsPart1 from "../assets/yugtun-audio/instructions-intro.mp3";
-import InstructionsPart2 from "../assets/yugtun-audio/instructions-running.mp3";
-import InstructionsPart3 from "../assets/yugtun-audio/instructions-analyzing.mp3";
+import InstructionsPart1 from "./assets/yugtun-audio/instructions-intro.mp3";
+import InstructionsPart2 from "./assets/yugtun-audio/instructions-running.mp3";
+import InstructionsPart3 from "./assets/yugtun-audio/instructions-analyzing.mp3";
 
-import PlayButton from "../assets/snapshots-directions/instructions-play-button@3x.png";
-import NewTrialButton from "../assets/snapshots-directions/instructions-new-trial-button@3x.png";
+import PlayButton from "./assets/snapshots-directions/instructions-play-button@3x.png";
+import NewTrialButton from "./assets/snapshots-directions/instructions-new-trial-button@3x.png";
 
 const instructionsPart1_text = `Experiment with a plant in a terrarium to see how much sugar it creates and uses
 with different conditions of light, water, and CO<sub>2</sub>.`;
